Update romcalc.asm
I moved the compare for the + sign into the separate method so that we could simply use it to separate the + from anything else. I also added in the start method mov dl, al so we could have 2 counters: one for loop1 and on for separate. In separate I think I did something right but just change it if I 'm not on the right track