Readability, finding plus char
Renamed labels for clarity, added code iterates through stored values,
finds the plus sign, and stores its offset in CX. Just now (while
typing) realized that is unnecessary, but the code can be reused.
Instead of storing the offset, we could be storing the converted
numerals in a register up until the plus is reached, then move to
another register or save and reuse that register for the next numeral. I
will most likely make that change tomorrow unless you get to it first.