Fixed!
Program now outputs reversed input. Some thoughts:
1.) Program still trims spaces even if quotes are used on the command
line, but the quotes appear in the reversed string. They might not
affect the math, not sure yet, so may/may not need to do something.
2.) We can't output values larger that 3999 as per the requirements, but
our program so far has no issue with that. I'm thinking we'll just do
the math as normal, and if value is greater than that, jump to a routine
that outputs an error message. That way, we don't need to check in the
beginning.