Still not summing correctly
Code wouldn't compile when CH was used, changed to CX. Decremented AL
early on so that the carriage return wouldn't be a factor. Added
firstChar routine so that CX would get an initial value, otherwise the
first check fails. Might not be the best way to solve it. The code still
isn't getting a correct answer. We need to fix that, and outputting var3
(which is in AX near the end) instead of the reversed input would help
speed up that debugging.